requirements/lint.txt
flake8==5.0.4
flake8-commas==2.1.0
flake8-debugger==4.1.2
flake8-django==1.1.5
flake8-isort==4.2.0
flake8-print==5.0.0
pep8-naming==0.13.2
